Export test scritp and create runner
I had to manually remove a line that called a magic function in the
python script that came out of nbconvert.

The runner calls from subprocess just to make sure the script actually
runs through with no errors.

Pushing this to my fork but expecting a build failure due to missing
dependencies.